Move GlobalOutbound into tools/execute; read OAuth form bodies as formData
GlobalOutbound is the outbound proxy for the execute isolate and its only
caller is tools/execute.ts (exports.GlobalOutbound(...)), so the class now
lives there next to runExecute. wrangler resolves the GLOBAL_OUTBOUND
worker-loader entrypoint from the worker entry module, so index.ts
re-exports it. index.ts drops the WorkerEntrypoint/fetchWithRetry imports.

Also silence the workerd '.text() on application/x-www-form-urlencoded'
warning in two OAuth tests by reading the mocked request body with
.formData() (the content-type-correct API) and asserting on parsed params
instead of substring-matching the raw body.
